PUTAIN DE BRESENHAM

+++ src/create_tab.c
@@ -6,14 +6,14 @@
 /*   By: nrivoire <nrivoire@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/06/25 17:51:37 by nrivoire     #+#   ##    ##    #+#       */
-/*   Updated: 2019/07/05 04:35:42 by nrivoire    ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/07/06 05:33:07 by nrivoire    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
 #include "fdf.h"
 
-t_point			create_point(t_lst *lst, float x, float y)
+t_point			create_point(t_lst *lst, int x, int y)
 {
 	t_point		data;
 
@@ -28,8 +28,8 @@ void			create_struct_tab(t_lst *lst, t_env *v)
 {
 	int			max;
 	int			inc;
-	float		x;
-	float		y;
+	int			x;
+	int			y;
 	t_lst		*tmp;
 
 	tmp = lst;
